Separate into header files and implement init routine
