+++ CMakeLists.txt
@@ -35,5 +35,3 @@ include(${C74_MIN_API_DIR}/script/min-posttarget.cmake)
 #############################################################
 # UNIT TEST
 #############################################################
-
-include(${C74_MIN_API_DIR}/test/min-object-unittest.cmake)
